Wavw for and sound quality

+++ app/src/main/java/com/instamelody/instamelody/utils/AudioDataReceivedListener.java
@@ -14,6 +14,8 @@
 
 package com.instamelody.instamelody.utils;
 
+
+
 public interface AudioDataReceivedListener {
-    void onAudioDataReceived(short[] data);
+    void onAudioDataReceived(byte[] data);
 }
